Sort buildoptions and move cli/build functions to internal
